requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
openai>=1.30.0
streamlit>=1.35.0
streamlit-agraph>=0.0.45
pyvis>=0.3.2
networkx>=3.3
pydantic>=2.0.0
python-dotenv>=1.0.0
azure-search-documents>=11.4.0
azure-core>=1.30.0
sqlalchemy>=2.0.0
pandas>=2.0.0
plotly>=5.20.0
graphviz>=0.20.1
chromadb>=0.5.0
tiktoken>=0.7.0
